backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.20
langchain-openai>=0.0.6
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.11
openai>=1.12.0
anthropic>=0.16.0
google-generativeai>=0.4.1
python-dotenv==1.0.0
httpx==0.26.0
pyyaml==6.0.1
# FastMCP for Model Context Protocol integration (REQUIRED)
fastmcp>=0.1.0